all 3 function working properly, creating delete and other exception and field level validation

+++ src/main/java/com/banking/accountmanagementapps/service/TransactionService.java
@@ -8,5 +8,5 @@ public interface TransactionService {
 
     TransactionDTO deposit(String accountNumber, BigDecimal amount);
     TransactionDTO withdrawn(String accountNumber, BigDecimal amount);
-    TransactionDTO transfer (String fromAccountNumber, String toAccountNumber, Double amount);
+    TransactionDTO transfer (String fromAccountNumber, String toAccountNumber, BigDecimal amount);
 }
